Split up tr_collect
