+++ spt.vcxproj
@@ -1258,6 +1258,7 @@ del "$(OutDir)spt-version.obj"</Command>
     <ClCompile Include="spt\features\visualizations\player_trace\tr_feature.cpp" />
     <ClCompile Include="spt\features\visualizations\player_trace\tr_record_cache.cpp" />
     <ClCompile Include="spt\features\visualizations\player_trace\tr_render_cache.cpp" />
+    <ClCompile Include="spt\features\visualizations\player_trace\tr_state_access.cpp" />
     <ClCompile Include="spt\features\visualizations\portal_placement.cpp" />
     <ClCompile Include="spt\features\visualizations\renderer\internal\materials_manager.cpp" />
     <ClCompile Include="spt\features\visualizations\renderer\internal\mesh_builder.cpp" />
