+++ spt/features/visualizations/player_trace/tr_render_cache.hpp
@@ -88,7 +88,7 @@ namespace player_trace
 				};
 
 				std::unordered_map<TrIdx<TrPhysMesh>, TrackedMesh> physObjs;
-				bool anyStale = false;
+				bool anyStale = true;
 			} ents;
 
 		} meshes;
